my_agent/requirements.txt
# Cloud dependencies for my_agent (Vertex AI Agent Engine)
google-adk>=0.4.0
google-cloud-aiplatform>=1.60.0
google-genai>=0.8.0
